# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: ID PASS DataCollect Backend API
  description: |
    REST API for the ID PASS DataCollect backend server. The server is the central
    sync hub for offline-first DataCollect clients (mobile, web, admin UI) and
    coordinates external integrations via the adapter registry.

    ## Authentication

    Most endpoints require a JWT access token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Access tokens are issued by `POST /api/users/login` and expire after one hour.
    Use `POST /api/users/refresh` to get a new access token (and rotated refresh token)
    without re-prompting for credentials — field agents use this to stay signed in
    while offline. Self-service endpoints (`/api/auth/**`) issue separate scoped
    tokens via OTP, national ID, or OIDC exchange.

    ## Multi-tenant (multi-program) support

    All tenant-scoped endpoints accept a `configId` query parameter (or field in
    the request body). Each configuration has its own entity forms, entity data,
    auth providers, and external-sync settings. Non-admin users can only access
    configurations listed in their `tenantIds` claim — the `validateTenantAccess`
    middleware enforces this on every tenant-scoped route.

    ## Synchronization

    The sync endpoints support bidirectional synchronization between clients and
    server, as well as external system integration. Internal sync uses cursor-based
...
    persisted to `sync_events`.
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin and user JWT issued by `/api/users/login` or `/api/users/refresh`
    SelfServiceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-service JWT issued by `/api/auth/otp/verify`, `/api/auth/id/verify`, or `/api/auth/oidc/exchange`
  parameters:
    ConfigIdQuery:
      name: configId
      in: query
      required: false
      schema:
        type: string
        default: default
      description: Program (tenant) identifier. Defaults to `default`.
  schemas:
...
    Event:
      type: object
      required: [guid, entityGuid, type, data, timestamp, userId, syncLevel]
      properties:
        guid:
          type: string
          format: uuid
        entityGuid:
          type: string
          description: Target entity GUID (UUID or human-readable ID such as `hh-001`)
        type:
          type: string
          description: Event type (e.g. `create-group`, `update-individual`, `add-member`)
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        userId:
          type: string
        syncLevel:
          type: integer
          enum: [0, 1]
          description: 0 = LOCAL, 1 = SYNCED
        schemaVersion:
          type: integer
    AuditLog:
      type: object
      required: [guid, timestamp, action, eventGuid, entityGuid, changes, signature]
      properties:
        guid:
          type: string
        timestamp:
          type: string
          format: date-time
        userId:
          type: string
        action:
          type: string
        eventGuid:
          type: string
        entityGuid:
          type: string
        changes:
          type: object
          additionalProperties: true
        signature:
          type: string
          description: Hash-chain signature for tamper detection
...
    SyncJob:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        configId: { type: string }
        status:
          type: string
          enum: [pending, success, partial, failed]
        phase:
          type: string
          enum: [pending, pushing, pulling, completed, failed, cancelled]
        pushed: { type: integer }
        pulled: { type: integer }
        failed: { type: integer }
        skipped: { type: integer }
        durationMs: { type: integer }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        triggeredBy: { type: string }
        errorMessage: { type: string, nullable: true }
        errors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              entityGuid: { type: string }
              code: { type: string }
              message: { type: string }
...
paths:
...
  /api/sync/pull:
    get:
      tags: [Synchronization]
      summary: Pull events from server
      description: |
        Returns events updated after `since`, paginated with a composite cursor. Pages are 10
        events by default (100 when `areaIds` is provided). When `areaIds` is set, events are
        server-side filtered to only those targeting entities in the listed areas.

        Unresolved potential duplicates do not block delivery but are surfaced as a `warnings`
        array in the response.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
        - name: since
          in: query
          schema: { type: string }
          description: Composite cursor `timestamp|eventGuid` or ISO timestamp. Defaults to epoch when absent.
        - name: areaIds
          in: query
          schema: { type: string }
          description: Comma-separated area IDs for selective sync
      responses:
        '200':
          description: Page of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/Event' }
                  nextCursor: { type: string, nullable: true }
                  warnings:
                    type: array
                    items: { type: string }
        '404':
          description: Tenant not found
  /api/sync/pull/callback:
    get:
      tags: [Synchronization]
      summary: External pull callback (reserved)
      description: Placeholder for async external-pull callbacks. Currently returns `not implemented`.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Placeholder response
  /api/sync/push:
    post:
      tags: [Synchronization]
      summary: Push events to server
      description: |
        Submits a batch of events for the tenant. When a direct postgres URL is available, the
        batch runs transactionally — all events commit or none are applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                configId: { type: string, default: default }
                events:
                  type: array
                  items: { $ref: '#/components/schemas/Event' }
      responses:
        '200':
          description: Batch applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: success }
                  applied: { type: integer }
                  failed: { type: integer }
                  errors:
                    type: array
                    items: { type: object, additionalProperties: true }
        '400':
          description: Invalid payload
        '404':
          description: Tenant not found
        '422':
          description: Batch rolled back; no events applied
  /api/sync/pull/audit-logs:
    get:
      tags: [Synchronization]
      summary: Pull audit logs
      description: Returns all audit log entries created after `since`.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
        - name: since
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditLog' }
        '404':
          description: Tenant not found
  /api/sync/push/audit-logs:
    post:
      tags: [Synchronization]
      summary: Push audit logs
      description: |
        Upload audit log entries for hash-chain integrity tracking. Each entry must carry a
        signature produced by the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [auditLogs]
              properties:
                configId: { type: string, default: default }
                auditLogs:
                  type: array
                  items: { $ref: '#/components/schemas/AuditLog' }
      responses:
        '200':
          description: Saved
        '400':
          description: Invalid payload
        '404':
          description: Tenant not found
        '500':
          description: Failed to save audit logs
  /api/sync/status:
    get:
      tags: [External Sync Jobs]
      summary: Get latest external sync status
      description: Returns `isSyncing`, the last completed job, and the currently active job (if any) for the given config.
      parameters:
        - name: configId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  isSyncing: { type: boolean }
                  lastEvent:
                    type: object
                    nullable: true
                    allOf:
                      - $ref: '#/components/schemas/SyncJob'
                  activeJob:
                    type: object
                    nullable: true
                    allOf:
                      - $ref: '#/components/schemas/SyncJob'
        '400':
          description: Missing configId
  /api/sync/events:
    get:
      tags: [External Sync Jobs]
      summary: List recent sync jobs
      description: Returns the 20 most recent sync events for the given config, newest first.
      parameters:
        - name: configId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Recent jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/SyncJob' }
        '400':
          description: Missing configId
  /api/sync/external:
    post:
      tags: [External Sync Jobs]
      summary: Trigger external sync
      description: |
        Starts a background sync job against the configured external system. Returns 202 with a
        `jobId` immediately — poll `/api/sync/external/{jobId}` or `/api/sync/status` for
        progress. Returns 409 if a job is already in progress for the same config.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                configId: { type: string, default: default }
                credentials:
                  type: object
                  properties:
                    username: { type: string }
                    password: { type: string }
                  description: Optional — only adapters that use basic auth require this
      responses:
        '202':
          description: Job scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string, format: uuid }
                  status: { type: string, example: pending }
        '400':
          description: Invalid payload
        '404':
          description: Tenant not found
        '409':
          description: Sync already in progress
        '503':
          description: Sync event store unavailable
  /api/sync/external/{jobId}:
    get:
      tags: [External Sync Jobs]
      summary: Get sync job by ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncJob' }
        '403':
          description: No access to the job's tenant
        '404':
          description: Job not found
        '503':
          description: Sync event store unavailable
  /api/sync/external/{jobId}/cancel:
    post:
      tags: [External Sync Jobs]
      summary: Cancel a running sync job
      description: Signals the running job's abort controller. The job transitions to `cancelled` phase.
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Cancellation requested
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: cancelling }
        '403':
          description: No access to the job's tenant
        '404':
          description: No active job with that ID
  /api/sync/external/{jobId}/retry:
    post:
      tags: [External Sync Jobs]
      summary: Retry a completed sync job
      description: Creates a new job using the same `configId` as the original. Returns 409 if another job is already running.
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: New job scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string, format: uuid }
                  status: { type: string, example: pending }
        '404':
          description: Original job not found or tenant not found
        '409':
          description: A sync is already in progress
        '503':
          description: Sync event store unavailable
...
security:
  - BearerAuth: []